app/apps/translate/schemas.py
from fastapi_mongo_base.schemas import UserOwnedEntitySchema
from fastapi_mongo_base.tasks import TaskMixin
from pydantic import BaseModel


class TranslateSchemaCreate(BaseModel):
    text: str
    user_id: str | None = None


class TranslateSchema(UserOwnedEntitySchema, TaskMixin, TranslateSchemaCreate):  # type: ignore[misc]
    result: str | None = None
    usage_amount: float | None = None
    usage_id: str | None = None
